+++ actions/vmwarelib/serialize.py
@@ -80,6 +80,11 @@ HOST_GET_NON_JSON_SERILIZABLE_TYPES = [
     vim.host.SystemIdentificationInfo,
 ]
 
+DATASTORE_GET_NON_JSON_SERILIZABLE_TYPES = [
+    vim.Datastore,
+    vim.Datastore.Summary
+]
+
 
 class MyJSONEncoder(json.JSONEncoder):
     def default(self, obj):  # pylint: disable=E0202
@@ -99,6 +104,18 @@ class HostGetJSONEncoder(json.JSONEncoder):
             return super(HostGetJSONEncoder, self).default(obj)
         except TypeError:
             if type(obj) in HOST_GET_NON_JSON_SERILIZABLE_TYPES:
+                return obj.__dict__
+
+            # For anything else just return the class name
+            return "__{}__".format(obj.__class__.__name__)
+
+
+class DatastoreGetJSONEncoder(json.JSONEncoder):
+    def default(self, obj):  # pylint: disable=E0202
+        try:
+            return super(DatastoreGetJSONEncoder, self).default(obj)
+        except TypeError:
+            if type(obj) in DATASTORE_GET_NON_JSON_SERILIZABLE_TYPES:
                 return obj.__dict__
 
             # For anything else just return the class name